Fix fqcn[deep] warnings
