+++ tests/unit/modules/network/nxos/test_nxos_zone_zoneset.py
@@ -8,10 +8,10 @@ from __future__ import absolute_import, division, print_function
 
 __metaclass__ = type
 
-from ansible_collections.cisco.nxos.plugins.modules.storage import nxos_zone_zoneset
+from ansible_collections.cisco.nxos.plugins.modules import nxos_zone_zoneset
 from ansible_collections.cisco.nxos.tests.unit.compat.mock import patch
 
-from ..nxos_module import TestNxosModule, load_fixture, set_module_args
+from .nxos_module import TestNxosModule, load_fixture, set_module_args
 
 
 class TestNxosZoneZonesetModule(TestNxosModule):
@@ -19,7 +19,7 @@ class TestNxosZoneZonesetModule(TestNxosModule):
 
     def setUp(self):
         super(TestNxosZoneZonesetModule, self).setUp()
-        module_path = "ansible_collections.cisco.nxos.plugins.modules.storage.nxos_zone_zoneset."
+        module_path = "ansible_collections.cisco.nxos.plugins.modules.nxos_zone_zoneset."
 
         self.mock_run_commands = patch(module_path + "run_commands")
         self.run_commands = self.mock_run_commands.start()
@@ -34,7 +34,9 @@ class TestNxosZoneZonesetModule(TestNxosModule):
         )
         self.execute_show_cmd_zoneset = self.mock_execute_show_cmd_zoneset.start()
 
-        self.mock_execute_show_cmd_zone = patch(module_path + "ShowZone.execute_show_zone_vsan_cmd")
+        self.mock_execute_show_cmd_zone = patch(
+            module_path + "ShowZone.execute_show_zone_vsan_cmd",
+        )
         self.execute_show_cmd_zone = self.mock_execute_show_cmd_zone.start()
 
         self.mock_execute_show_cmd_zone_status = patch(
@@ -213,13 +215,20 @@ class TestNxosZoneZonesetModule(TestNxosModule):
 
     # Test zone  add/removal
     def test_zone_add_rem(self):
-        a = dict(zone_zoneset_details=[dict(vsan=923, zone=[dict(name="zoneB", remove=True)])])
+        a = dict(
+            zone_zoneset_details=[
+                dict(vsan=923, zone=[dict(name="zoneB", remove=True)]),
+            ],
+        )
         set_module_args(a, True)
         self.execute_show_cmd_zone_status.return_value = load_fixture(
             "nxos_zone_zoneset",
             "shzonestatus_1.cfg",
         )
-        self.execute_show_cmd_zone.return_value = load_fixture("nxos_zone_zoneset", "shzone_0.cfg")
+        self.execute_show_cmd_zone.return_value = load_fixture(
+            "nxos_zone_zoneset",
+            "shzone_0.cfg",
+        )
         result = self.execute_module(changed=True)
         self.assertEqual(
             result["commands"],
@@ -232,13 +241,20 @@ class TestNxosZoneZonesetModule(TestNxosModule):
         )
 
     def test_zone_add_rem_1(self):
-        a = dict(zone_zoneset_details=[dict(vsan=923, zone=[dict(name="zoneC", remove=True)])])
+        a = dict(
+            zone_zoneset_details=[
+                dict(vsan=923, zone=[dict(name="zoneC", remove=True)]),
+            ],
+        )
         set_module_args(a, True)
         self.execute_show_cmd_zone_status.return_value = load_fixture(
             "nxos_zone_zoneset",
             "shzonestatus_1.cfg",
         )
-        self.execute_show_cmd_zone.return_value = load_fixture("nxos_zone_zoneset", "shzone_0.cfg")
+        self.execute_show_cmd_zone.return_value = load_fixture(
+            "nxos_zone_zoneset",
+            "shzone_0.cfg",
+        )
         result = self.execute_module(changed=False, failed=False)
         m = "zone 'zoneC' is not present in vsan 923"
         assert m in str(result["messages"])
@@ -251,7 +267,10 @@ class TestNxosZoneZonesetModule(TestNxosModule):
             "nxos_zone_zoneset",
             "shzonestatus_1.cfg",
         )
-        self.execute_show_cmd_zone.return_value = load_fixture("nxos_zone_zoneset", "shzone_0.cfg")
+        self.execute_show_cmd_zone.return_value = load_fixture(
+            "nxos_zone_zoneset",
+            "shzone_0.cfg",
+        )
         result = self.execute_module(changed=True)
         self.assertEqual(
             result["commands"],
@@ -270,7 +289,10 @@ class TestNxosZoneZonesetModule(TestNxosModule):
             "nxos_zone_zoneset",
             "shzonestatus_1.cfg",
         )
-        self.execute_show_cmd_zone.return_value = load_fixture("nxos_zone_zoneset", "shzone_0.cfg")
+        self.execute_show_cmd_zone.return_value = load_fixture(
+            "nxos_zone_zoneset",
+            "shzone_0.cfg",
+        )
         result = self.execute_module(changed=False, failed=False)
         m = "zone 'zoneB' is already present in vsan 923"
         assert m in str(result["messages"])
@@ -293,7 +315,10 @@ class TestNxosZoneZonesetModule(TestNxosModule):
             "nxos_zone_zoneset",
             "shzonestatus_1.cfg",
         )
-        self.execute_show_cmd_zone.return_value = load_fixture("nxos_zone_zoneset", "shzone_0.cfg")
+        self.execute_show_cmd_zone.return_value = load_fixture(
+            "nxos_zone_zoneset",
+            "shzone_0.cfg",
+        )
         result = self.execute_module(changed=True)
         self.assertEqual(
             result["commands"],
@@ -312,14 +337,19 @@ class TestNxosZoneZonesetModule(TestNxosModule):
         mem1 = {"pwwn": "11:11:11:11:11:11:11:11", "remove": True}
         mem2 = {"device_alias": "test123", "remove": True}
         a = dict(
-            zone_zoneset_details=[dict(vsan=923, zone=[dict(name="zoneA", members=[mem1, mem2])])],
+            zone_zoneset_details=[
+                dict(vsan=923, zone=[dict(name="zoneA", members=[mem1, mem2])]),
+            ],
         )
         set_module_args(a, True)
         self.execute_show_cmd_zone_status.return_value = load_fixture(
             "nxos_zone_zoneset",
             "shzonestatus_1.cfg",
         )
-        self.execute_show_cmd_zone.return_value = load_fixture("nxos_zone_zoneset", "shzone_0.cfg")
+        self.execute_show_cmd_zone.return_value = load_fixture(
+            "nxos_zone_zoneset",
+            "shzone_0.cfg",
+        )
         result = self.execute_module(changed=True)
         self.assertEqual(
             result["commands"],
@@ -338,14 +368,19 @@ class TestNxosZoneZonesetModule(TestNxosModule):
         mem1 = {"pwwn": "11:11:11:11:11:11:11:11", "remove": True}
         mem2 = {"device_alias": "test123", "remove": True}
         a = dict(
-            zone_zoneset_details=[dict(vsan=923, zone=[dict(name="zoneA1", members=[mem1, mem2])])],
+            zone_zoneset_details=[
+                dict(vsan=923, zone=[dict(name="zoneA1", members=[mem1, mem2])]),
+            ],
         )
         set_module_args(a, True)
         self.execute_show_cmd_zone_status.return_value = load_fixture(
             "nxos_zone_zoneset",
             "shzonestatus_1.cfg",
         )
-        self.execute_show_cmd_zone.return_value = load_fixture("nxos_zone_zoneset", "shzone_0.cfg")
+        self.execute_show_cmd_zone.return_value = load_fixture(
+            "nxos_zone_zoneset",
+            "shzone_0.cfg",
+        )
         result = self.execute_module(changed=False, failed=False)
         m = "zone 'zoneA1' is not present in vsan 923 , hence cannot remove the members"
         assert m in str(result["messages"])
@@ -369,7 +404,10 @@ class TestNxosZoneZonesetModule(TestNxosModule):
             "nxos_zone_zoneset",
             "shzonestatus_0.cfg",
         )
-        self.execute_show_cmd_zone.return_value = load_fixture("nxos_zone_zoneset", "shzone_1.cfg")
+        self.execute_show_cmd_zone.return_value = load_fixture(
+            "nxos_zone_zoneset",
+            "shzone_1.cfg",
+        )
         result = self.execute_module(changed=True)
         self.assertEqual(
             result["commands"],
@@ -387,13 +425,20 @@ class TestNxosZoneZonesetModule(TestNxosModule):
     def test_zonemem_add_rem_4(self):
         mem2 = {"device_alias": "test123", "devtype": "both", "remove": True}
 
-        a = dict(zone_zoneset_details=[dict(vsan=922, zone=[dict(name="zoneA", members=[mem2])])])
+        a = dict(
+            zone_zoneset_details=[
+                dict(vsan=922, zone=[dict(name="zoneA", members=[mem2])]),
+            ],
+        )
         set_module_args(a, True)
         self.execute_show_cmd_zone_status.return_value = load_fixture(
             "nxos_zone_zoneset",
             "shzonestatus_0.cfg",
         )
-        self.execute_show_cmd_zone.return_value = load_fixture("nxos_zone_zoneset", "shzone_1.cfg")
+        self.execute_show_cmd_zone.return_value = load_fixture(
+            "nxos_zone_zoneset",
+            "shzone_1.cfg",
+        )
         result = self.execute_module(changed=True)
         self.assertEqual(
             result["commands"],
@@ -408,7 +453,9 @@ class TestNxosZoneZonesetModule(TestNxosModule):
     # Test zoneset add/removal
     def test_zoneset_add_rem(self):
         a = dict(
-            zone_zoneset_details=[dict(vsan=922, zoneset=[dict(name="zsetname21", remove=True)])],
+            zone_zoneset_details=[
+                dict(vsan=922, zoneset=[dict(name="zsetname21", remove=True)]),
+            ],
         )
         set_module_args(a, True)
         self.execute_show_cmd_zone_status.return_value = load_fixture(
@@ -430,7 +477,9 @@ class TestNxosZoneZonesetModule(TestNxosModule):
         )
 
     def test_zoneset_add_rem_1(self):
-        a = dict(zone_zoneset_details=[dict(vsan=922, zoneset=[dict(name="zsetname21New")])])
+        a = dict(
+            zone_zoneset_details=[dict(vsan=922, zoneset=[dict(name="zsetname21New")])],
+        )
         set_module_args(a, True)
         self.execute_show_cmd_zone_status.return_value = load_fixture(
             "nxos_zone_zoneset",
@@ -451,7 +500,9 @@ class TestNxosZoneZonesetModule(TestNxosModule):
         )
 
     def test_zoneset_add_rem_2(self):
-        a = dict(zone_zoneset_details=[dict(vsan=922, zoneset=[dict(name="zsetname21")])])
+        a = dict(
+            zone_zoneset_details=[dict(vsan=922, zoneset=[dict(name="zsetname21")])],
+        )
         set_module_args(a, True)
         self.execute_show_cmd_zone_status.return_value = load_fixture(
             "nxos_zone_zoneset",
@@ -570,7 +621,9 @@ class TestNxosZoneZonesetModule(TestNxosModule):
     # Test zoneset activate/deactivate
     def test_zoneset_activate_deactivate(self):
         a = dict(
-            zone_zoneset_details=[dict(vsan=221, zoneset=[dict(name="zsv221", action="activate")])],
+            zone_zoneset_details=[
+                dict(vsan=221, zoneset=[dict(name="zsv221", action="activate")]),
+            ],
         )
         set_module_args(a, True)
         self.execute_show_cmd_zone_status.return_value = load_fixture(
@@ -659,14 +712,19 @@ class TestNxosZoneZonesetModule(TestNxosModule):
         mem1 = {"pwwn": "21:01:00:1b:32:a1:c0:a8", "remove": True}
         mem2 = {"pwwn": "50:06:01:6a:47:e4:6e:59", "remove": True}
         a = dict(
-            zone_zoneset_details=[dict(vsan=221, zone=[dict(name="zv221", members=[mem1, mem2])])],
+            zone_zoneset_details=[
+                dict(vsan=221, zone=[dict(name="zv221", members=[mem1, mem2])]),
+            ],
         )
         set_module_args(a, True)
         self.execute_show_cmd_zone_status.return_value = load_fixture(
             "nxos_zone_zoneset",
             "shzonestatus_4.cfg",
         )
-        self.execute_show_cmd_zone.return_value = load_fixture("nxos_zone_zoneset", "shzone_2.cfg")
+        self.execute_show_cmd_zone.return_value = load_fixture(
+            "nxos_zone_zoneset",
+            "shzone_2.cfg",
+        )
         self.execute_show_cmd_zoneset_active.return_value = load_fixture(
             "nxos_zone_zoneset",
             "shzonesetactive_0.cfg",
